[CIDMDEV-4973] - Fix all Cloud20IntegrationTest test errors

+++ src/main/java/com/rackspace/idm/domain/entity/PaginatorContext.java
@@ -22,7 +22,7 @@ public class PaginatorContext<T> {
 
     private int offset;
     private int limit;
-    private int totalRecords;
+    private long totalRecords;
 
 
     public List<SearchResultEntry> getSearchResultEntryList() {
